% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyLoad.R
\name{easy.load}
\alias{easy.load}
\title{An even easier way to import your data!}
\usage{
easy.load()
}
\value{
A complete line of code ready to be copy/pasted into
  the command line or R script
}
\description{
\code{easy.load} generates the code required to import your data by
  asking questions from the R console
}
\details{
\code{easy.load} will begin the import wizard. A series of
  questions presented from the R console will ask for the necessary
  information to import your data in a user-friendly way. See
  \code{\link{load.data}} for parameter definitions and conventions
  for how to answer \code{easy.load} questions if you get stuck.

  \code{easy.load} uses an interactive window to locate files
  to create the path name. The window usually appears BEHIND
  the R session and other open programs, so minimize open
  windows until the interactive pop-up window is visible. After
  you browse to your file location, bring the R session window
  back up to continue.

 Follow any instructions accompanying the questions exactly,
 as some prompts require a very specific answer format.
}
\examples{
\dontrun{
# example output from easy.load (after answering questions)
myData = load.data('/Users/Username/Folder/', 'Positive', 20, -200, 899, 'n', FALSE)
}
}
\author{
Travis Moore
}

